Preparing for 0.0.6 release
Only one bug remains in the code, and that is the edge case wherein strong or em tags start in a quotation and end outside the quotation or start outside the quotation and end inside. While neither should technically happen, it should be able to function despite that.

Also, perhaps add a feature to change what style of double quote (e.g., normal English double quotes, double chevrons in, double chevrons out, etc.)